Made webpack more modular
