+++ config/webpack/helpers/get-absolute-path.js
@@ -1,0 +1,6 @@
+const path = require('path');
+
+const getAbsolutePath = relativePath =>
+  path.join(__dirname, '../', relativePath);
+
+module.exports.getAbsolutePath = getAbsolutePath;
