+++ config/webpack/helpers/get-app-routes.js
@@ -1,0 +1,9 @@
+const { getAppManifests } = require('../helpers/get-app-manifests');
+
+function getAppRoutes() {
+  return getAppManifests()
+    .map(m => m.rootUrl)
+    .filter(m => m);
+}
+
+module.exports.getAppRoutes = getAppRoutes;
