+++ package.json
@@ -8,12 +8,12 @@
     "build": "./script/build.sh --scaffold",
     "build-analyze": "yarn build:production --analyzer; npm run analyze",
     "build:production": "NODE_ENV=production yarn build --buildtype=vagovprod",
-    "build:webpack": "NODE_OPTIONS=--max-old-space-size=4096 webpack --config config/webpack.config.js",
+    "build:webpack": "NODE_OPTIONS=--max-old-space-size=4096 webpack --config config/webpack/webpack.config.js",
     "build:webpack:prod": "NODE_ENV=production yarn build:webpack --env.buildtype vagovprod",
     "build:webpack:staging": "yarn build:webpack --env.buildtype vagovstaging",
     "build:webpack:dev": "yarn build:webpack --env.buildtype vagovdev",
     "build:webpack:local": "yarn build:webpack --env.buildtype localhost",
-    "debug:webpack:prod": "NODE_ENV=production NODE_OPTIONS=--max-old-space-size=4096 node --inspect --inspect-brk ./node_modules/webpack/bin/webpack.js --config config/webpack.config.js --env.buildtype vagovprod",
+    "debug:webpack:prod": "NODE_ENV=production NODE_OPTIONS=--max-old-space-size=4096 node --inspect --inspect-brk ./node_modules/webpack/bin/webpack.js --config config/webpack/webpack.config.js --env.buildtype vagovprod",
     "cy:open": "cypress open --config-file config/cypress.json",
     "cy:run": "cypress run --config-file config/cypress.json",
     "cy:test:docker": "node script/run-cypress-tests-docker.js",
